+++ spec/support/test_strings.rb
@@ -1,0 +1,13 @@
+def test_strings
+  {
+    "Hello World" => "hello-world",
+    "one Flew_South" => "one-flew_south",
+    "Attr: Halp" => "attr-halp",
+    "Freddie & Bananas" => "freddie-and-bananas",
+    "One + Two = Three" => "one-plus-two-equals-three",
+    " hello world" => "hello-world",
+    "hello world " => "hello-world",
+    "post id 1234" => "post-id-1234",
+    "124 post id" => "124-post-id"
+  }
+end